chore: mv recover to start cmd
